Revert "Version upgrade 4.0.2"

+++ iosApp/iosApp/Services/DataUploadManager.swift
@@ -29,21 +29,23 @@ class DataUploadManager {
                 self.observationDataRepository.allAsBulk { [weak self] dataBulk in
                     if let dataBulk, !dataBulk.dataPoints.isEmpty {
                         print("Sending data to backend...")
-                        AppDelegate.shared.networkService.iosSendData(data: dataBulk) { pair in
-                            if let error = pair.second {
-                                print("Error: \(error)")
-                                self?.currentlyUploading = false
-                                completion(false)
-                            } else if let self, let idSet = pair.first as? Set<String> {
-                                print("Sent data! Deleting local data...")
-                                self.observationDataRepository.deleteAllWithId(idSet: idSet)
-                                print("Deleted data!")
-                                self.currentlyUploading = false
-                                completion(true)
-                            } else {
-                                print("Error!")
-                                self?.currentlyUploading = false
-                                completion(false)
+                        AppDelegate.shared.networkService.iosSendData(data: dataBulk) { weakReference in
+                            if let pair = weakReference.value {
+                                if let error = pair.second {
+                                    print("Error: \(error)")
+                                    self?.currentlyUploading = false
+                                    completion(false)
+                                } else if let self, let idSet = pair.first as? Set<String> {
+                                    print("Sent data! Deleting local data...")
+                                    self.observationDataRepository.deleteAllWithId(idSet: idSet)
+                                    print("Deleted data!")
+                                    self.currentlyUploading = false
+                                    completion(true)
+                                } else {
+                                    print("Error!")
+                                    self?.currentlyUploading = false
+                                    completion(false)
+                                }
                             }
                         }
                     } else {
